refactor(web): delete unused UserAvatar Settings menu
Settings already lives in account-menu-config. Remove the unmounted UserAvatarClient stack and keep UserAvatarContent plus workspace-switcher.
