--- apps/web/src/app/(app)/components/user-avatar/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./user-avatar";
